+++ CHANGELOG.rst
@@ -25,14 +25,14 @@ Changes:
 ^^^^^^^^
 
 - ``structlog.processors.ExceptionPrettyPrinter`` and ``structlog.processors.format_exc_info`` now support passing of Exceptions on Python 3.
+- Clean up the context when exiting ``structlog.threadlocal.tmp_bind`` in case of exceptions.
+  [`64 <https://github.com/hynek/structlog/issues/64>`_]
 - Be more more lenient about missing ``__name__``\ s.
   [`62 <https://github.com/hynek/structlog/pull/62>`_]
 - Add ``structlog.dev.ConsoleRenderer`` that renders the event dictionary aligned and with colors.
 - Use `six <https://pythonhosted.org/six/>`_ for compatibility.
 - Add ``structlog.processors.UnicodeDecoder`` that will decode all byte string values in an event dictionary to Unicode.
 - Add ``serializer`` parameter to ``structlog.processors.JSONRenderer`` which allows for using different (possibly faster) JSON encoders than the standard library.
-- Clean up the context when exiting ``structlog.threadlocal.tmp_bind`` in case of exceptions.
-  [`64 <https://github.com/hynek/structlog/issues/64>`_]
 
 
 15.3.0 (2015-09-25)